chore: CycleClock clean ups
1. Remove redundant CycleClock code that I have not cleaned up before
2. Expose `ThisFiber::GetRunningTimeCycles` to be able to check for how long the current fiber is running.
