+++ util/fibers/CMakeLists.txt
@@ -7,7 +7,7 @@ endif()
 
 add_library(fibers2 fibers.cc proactor_base.cc synchronization.cc
             fiber_file.cc epoll_proactor.cc epoll_socket.cc pool.cc
-            detail/scheduler.cc detail/fiber_interface.cc detail/wait_queue.cc detail/utils.cc
+            detail/scheduler.cc detail/fiber_interface.cc detail/wait_queue.cc
             accept_server.cc
             fiber_socket_base.cc listener_interface.cc
             prebuilt_asio.cc proactor_pool.cc stacktrace.cc
