+++ util/fibers/detail/fiber_interface.h
@@ -233,6 +233,8 @@ class FiberInterface {
     return tl.epoch;
   }
 
+  uint64_t GetRunningTimeCycles() const;
+
  protected:
   static constexpr uint16_t kTerminatedBit = 0x1;
   static constexpr uint16_t kBusyBit = 0x2;
